implemented degree access from outside

+++ src/main/java/com/signalcollect/sna/SignalCollectGephiConnector.java
@@ -2,11 +2,12 @@ package com.signalcollect.sna;
 
 import java.util.List;
 
+
 public interface SignalCollectGephiConnector {
 	
 	public void executeGraph();
 
-	public int getAverage();
+	public double getAverage();
 	
 	public List<Integer> getAll();
